tock-register-interface: move IntLike, RegisterLongName, ...
Move IntLike, RegisterLongName out of crate::registers to the
top-level crate and TryFromValue to crate::fields respectively. These
traits are independent and should therefore not be included in the
registers module. The reasoning for moving IntLike and
RegisterLongName to the top-level crate exports is that these are base
types over which all registers and fields are composed, and thus
represent essential functionality of the library. TryFromValue only
relates to individual register fields and is thus moved to
crate::fields.

Other code in the Tock repository is updated to reflect these changes.
