+++ chips/litex/src/litex_registers.rs
@@ -31,13 +31,13 @@
 //! }`, `BaseWidth` in `{ 32 }`.
 
 use core::marker::PhantomData;
-use tock_registers::fields::{Field, FieldValue};
+use tock_registers::fields::{Field, FieldValue, TryFromValue};
 use tock_registers::interfaces::{Readable, Writeable};
 pub use tock_registers::register_bitfields;
 use tock_registers::registers::{
-    IntLike as TRIntLike, LocalRegisterCopy, ReadOnly as TRReadOnly, ReadWrite as TRReadWrite,
-    RegisterLongName, TryFromValue, WriteOnly as TRWriteOnly,
+    LocalRegisterCopy, ReadOnly as TRReadOnly, ReadWrite as TRReadWrite, WriteOnly as TRWriteOnly,
 };
+use tock_registers::{IntLike as TRIntLike, RegisterLongName};
 
 /// Extension of the `tock_registers` `IntLike` trait.
 ///
