+++ kernel/src/common/mod.rs
@@ -12,8 +12,8 @@ pub mod registers {
     pub use tock_registers::interfaces;
     pub use tock_registers::registers::InMemoryRegister;
     pub use tock_registers::registers::LocalRegisterCopy;
-    pub use tock_registers::registers::RegisterLongName;
     pub use tock_registers::registers::{Aliased, ReadOnly, ReadWrite, WriteOnly};
+    pub use tock_registers::RegisterLongName;
     pub use tock_registers::{register_bitfields, register_structs};
 }
