+++ libraries/riscv-csr/src/csr.rs
@@ -4,7 +4,7 @@ use core::marker::PhantomData;
 
 use tock_registers::fields::Field;
 use tock_registers::interfaces::{Readable, Writeable};
-use tock_registers::registers::{IntLike, RegisterLongName};
+use tock_registers::{IntLike, RegisterLongName};
 
 pub const MINSTRETH: usize = 0xB82;
 pub const MINSTRET: usize = 0xB02;
